+++ src/components/layoutWrapper/LayoutWrapper.tsx
@@ -3,6 +3,7 @@
 import { usePathname } from 'next/navigation';
 import Header from '../header/header/Header';
 import BottomNavigationBar from '../bottomNavigationBar/bottomNavigationBar/BottomNavigationBar';
+import AuthGuard from '../authGuard/AuthGuard';
 
 export default function LayoutWrapper({ children }: { children: React.ReactNode }) {
   const pathname = usePathname();
@@ -12,7 +13,11 @@ export default function LayoutWrapper({ children }: { children: React.ReactNode 
     <div className="layout-container">
       <div className="mobile-frame">
         {!isLoginPage && <Header />}
-        <main>{children}</main>
+        
+        <main>
+          {isLoginPage ? children : <AuthGuard>{children}</AuthGuard>}
+        </main>
+        
         {!isLoginPage && <BottomNavigationBar />}
       </div>
     </div>
