--- src/hooks/useAuthRedirect.tsx
@@ -1,24 +1,0 @@
-'use client';
-
-import { useEffect, useState } from 'react';
-import { useRouter } from 'next/navigation';
-
-const useAuthRedirect = () => {
-  const router = useRouter();
-  const [isAuthenticated, setIsAuthenticated] = useState<boolean | null>(null);
-
-  useEffect(() => {
-    const accessToken = typeof window !== 'undefined' ? localStorage.getItem('accessToken') : null;
-
-    if (!accessToken) {
-      router.replace('/login');
-      setIsAuthenticated(false);
-    } else {
-      setIsAuthenticated(true);
-    }
-  }, [router]);
-
-  return { isAuthenticated };
-};
-
-export default useAuthRedirect;
